Moved the Camera and SlapLockWhirlWind classes to their own sets of files.
